US PhD gender diversity

This dataset contains the information pertaining to the gender diversity among the PhD graduates in the US. The data encompasses the number of graduates between the time period of 1950-2020.

us_phd_gender

Format

A data frame with 53 rows and 4 variables.

Year

The academic year of the gender diversity records

Male

The number of male PhD graduates (in 1000s) for the corresponding academic year

Female

The number of female PhD graduates (in 1000s) for the corresponding academic year

History

An indicator to classify whether the academic year is post or pre 2000

Source

https://www.statista.com/statistics/185167/number-of-doctoral-degrees-by-gender-since-1950/

Examples

head(us_phd_gender)
#> # A tibble: 6 × 4
#>   Year     Male Female History  
#>   <chr>   <dbl>  <dbl> <chr>    
#> 1 2019-20  85.2  105.  Post 2000
#> 2 2018-19  85.8  102.  Post 2000
#> 3 2017-18  85.4   98.4 Post 2000
#> 4 2016-17  84.6   96.7 Post 2000
#> 5 2015-16  84.2   93.9 Post 2000
#> 6 2014-15  84.9   93.6 Post 2000